\name{get_scriptdir}
\alias{get_scriptdir}
\title{
Directory of current script
}
\description{
Directory of current script
}
\usage{
get_scriptdir()
}
\details{
...
}
\value{
If the R script is not run from the command-line, it returns \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
