\name{get_scriptname}
\alias{get_scriptname}
\title{
File name of current script
}
\description{
File name of current script
}
\usage{
get_scriptname()
}
\details{
...
}
\value{
If the R script is not run from the command-line, it returns \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
