\name{source_script}
\alias{source_script}
\title{
Source the R script with command-line arguments
}
\description{
Source the R script with command-line arguments
}
\usage{
source_script(file, ..., argv_str = NULL)
}
\arguments{

  \item{file}{The R script}
  \item{...}{Pass to \code{\link[base]{source}}.}
  \item{argv_str}{The command-line arguments.}

}
\examples{
# There is no example
NULL

}
