\name{get_scriptname}
\alias{get_scriptname}
\title{
Full path of current script
}
\description{
Full path of current script
}
\usage{
get_scriptname()
}
\details{
...
}
\value{
If the R script is not run under command-line, it return \code{NULL}.
}
\author{
Zuguang Gu <z.gu@dkfz.de>
}
\examples{
# There is no example
NULL

}
