\name{GetOptions}
\alias{GetOptions}
\title{
Wrapper of the Perl module \code{Getopt::Long} in R
}
\description{
Wrapper of the Perl module \code{Getopt::Long} in R
}
\usage{
GetOptions(...)
}
\arguments{

  \item{...}{pass to \code{\link{GetoptLong}}}

}
\details{
This function is the same as \code{\link{GetoptLong}}. Just to make it consistent as the \code{GetOptions} 
subroutine in \code{Getopt::Long} module in Perl.
}
