\name{subCommands}
\alias{subCommands}
\title{
Setting sub commands
}
\description{
Setting sub commands
}
\usage{
subCommands(..., help_head = NULL, help_foot = NULL,
    argv_str = NULL)
}
\arguments{

  \item{...}{Specification of commands.}
  \item{help_head}{Head of the help message when invoking \code{Rscript foo.R --help}.}
  \item{help_foot}{Foot of the help message when invoking \code{Rscript foo.R --help}.}
  \item{argv_str}{A string that contains command-line arguments. It is only for testing purpose.}

}
\examples{
# There is no example
NULL
}
